### 2024-06-13 #151 @editorialbot
[PRE REVIEW]: JSOSuite.jl: Solving continuous optimization problems with JuliaSmoothOptimizers
**Submitting author:** @tmigot (<a href="http://orcid.org/0000-0001-7729-2513">Tangi Migot</a>)
**Repository:** https://github.com/JuliaSmoothOptimizers/JSOSuite.jl
**Branch with paper.md** (empty if default branch): jcon-paper
**Version:** 0.1.1
**Managing EiC:** Mathieu Besançon

Thanks for submitting your paper to JCON @tmigot. **Currently, there isn't a JCON editor assigned** to your paper.
