### 2024-06-13 8dec25e
Bump to OptimizationProblems 0.8 in tests (#168)
* Bump to OptimizationProblems 0.8 in tests

* Fix type argument
