+++ test/Project.toml
@@ -44,7 +44,7 @@ NLPModelsJuMP = "0.11, 0.12"
 NLPModelsKnitro = "0.7, 0.8, 0.9"
 NLPModelsModifiers = "0.6.2, 0.7"
 NLSProblems = "0.5"
-OptimizationProblems = "0.7.0 - 0.7.3"
+OptimizationProblems = "0.8"
 Percival = "0.6, 0.7"
 QuadraticModels = "0.9.2"
 RipQP = "0.5, 0.6"
