### 2024-06-13 e55f774
CompatHelper: bump compat for ADNLPModels to 0.8, (keep existing compat) (#167)
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
