### 2024-06-13 8365906
CompatHelper: bump compat for JSOSolvers to 0.12, (keep existing compat) (#164)
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
