### 2024-06-13 #151 @odow
@frapac or @sshin23 are either of you free to review this? JuliaCon functions like JOSS, so it shouldn't require too much time. The paper is only two pages, and once we have two reviewers, there is a short checklist of review items.
