### 2024-06-13 #151 @odow
@tmigot: you should take a look at the comments of the bot: a few DOIs need fixing, and you should include an explicit "Statement of need" section.

You can make changes by pushing new commits to your repo, and then commenting `@editorialbot generate pdf` in this issue.
