+++ paper/ref.bib
@@ -118,7 +118,7 @@ year="2006",
 publisher="Springer US",
 address="Boston, MA",
 pages="35--59",
-doi="10.1007/0-387-30065-1\_4",
+doi="10.1007/0-387-30065-1",
 }
 
 @software{montoison-orban-siquiera-nlpmodelsjump-2020,
@@ -382,5 +382,6 @@ year = {2023}
   title={Accelerating optimal power flow with {GPU}s: {SIMD} abstraction of nonlinear programs and condensed-space interior-point methods},
   author={Shin, Sungho and Pacaud, Fran{\c{c}}ois and Anitescu, Mihai},
   journal={arXiv preprint arXiv:2307.16830},
-  year={2023}
+  year={2023},
+  doi = {10.2139/ssrn.4601442}
 }
