+++ paper/paper.tex
@@ -29,7 +29,7 @@ where \(f:\mathbb{R}^n\to\mathbb{R}\) and  \(c:\mathbb{R}^n\to\mathbb{R}^m\) are
 Bounds on the variables appear separately from other types of constraints because numerical methods often treat them differently.
 %Equality constraints are modeled by setting components  \(-\infty < c_{L,i}=c_{U,i} < +\infty\) in~\eqref{eq:nlp}.
 
-\section{Optimization solvers and JuliaSmoothOptimizers}
+\section{Statement of Need}
 
 \texttt{JSOSuite.jl} is part of the JuliaSmoothOptimizers (JSO) ecosystem, % \cite{jso}.
 an academic organization that offers a collection of Julia packages for nonlinear optimization software development, testing, and benchmarking.
