### 2024-06-14 #151 @tmigot
> @tmigot: you should take a look at the comments of the bot: a few DOIs need fixing, and you should include an explicit "Statement of need" section.
>
> You can make changes by pushing new commits to your repo, and then commenting `@editorialbot generate pdf` in this issue.

Thanks @odow for the feedback. I missed that the "Statement of need" section title was mandatory, it is now corrected. DOI issues are fixed.
