### 2024-06-14 #151 @odow
@mzagorowska and @sshin23 thank you for agreeing to review.

The review will take place over at: https://github.com/JuliaCon/proceedings-review/issues/161

For the review, start by generating a checklist with `@editorialbot generate my checklist` and then check of each item in turn.

@mzagorowska: there is no explicit deadline, you're a volunteer, so it's just a best effort that fits your schedule.
