### 2024-06-15 #161 @odow
Thanks @sshin23!

@tmigot: You don't seem to cite ReverseDiff or Algencan in the paper? They can be removed from the .bib. I wonder if it also needs to be `Statement of need` 😆

I'm not sure why https://github.com/JuliaCon/proceedings-papers/blob/jcon.00161/jcon.00161/10.21105.jcon.00161.pdf
