### 2024-06-16 6801c41
Update to 0.6.0 and add CirrusCI
