### 2024-06-16 6b3de17
Update to 0.6.0 and add CirrusCI
