+++ test/runtests.jl
@@ -33,7 +33,7 @@ using SolverBenchmark
 meta = OptimizationProblems.meta
 
 function test_in_place_solve(nlp, solver_name)
-  pkg_name = JSOSuite.optimizers[JSOSuite.optimizers.name_solver.==solver_name, :name_pkg][1]
+  pkg_name = JSOSuite.optimizers[JSOSuite.optimizers.name_solver .== solver_name, :name_pkg][1]
   pkg_name = replace(pkg_name, ".jl" => "")
   solver = eval(Meta.parse(pkg_name * ".$solver_name"))(nlp)
   stats = solve!(solver, nlp)
@@ -46,7 +46,7 @@ end
 
 function test_in_place_solve(model::JuMP.Model, solver_name)
   nlp = MathOptNLPModel(model)
-  pkg_name = JSOSuite.optimizers[JSOSuite.optimizers.name_solver.==solver_name, :name_pkg][1]
+  pkg_name = JSOSuite.optimizers[JSOSuite.optimizers.name_solver .== solver_name, :name_pkg][1]
   pkg_name = replace(pkg_name, ".jl" => "")
   solver = eval(Meta.parse(pkg_name * ".$solver_name"))(nlp)
   stats = solve!(solver, model)
@@ -64,16 +64,16 @@ end
     solver_name == :DCIWorkspace && continue
     solver_name == :RipQPSolver && continue
     is_available =
-      JSOSuite.optimizers[JSOSuite.optimizers.name_solver.==solver_name, :is_available]
+      JSOSuite.optimizers[JSOSuite.optimizers.name_solver .== solver_name, :is_available]
     can_solve_nlp =
-      JSOSuite.optimizers[JSOSuite.optimizers.name_solver.==solver_name, :can_solve_nlp]
+      JSOSuite.optimizers[JSOSuite.optimizers.name_solver .== solver_name, :can_solve_nlp]
     spec_nls =
-      JSOSuite.optimizers[JSOSuite.optimizers.name_solver.==solver_name, :specialized_nls]
+      JSOSuite.optimizers[JSOSuite.optimizers.name_solver .== solver_name, :specialized_nls]
     if is_available[1] && can_solve_nlp[1]
       test_in_place_solve(nlp, solver_name)
       test_in_place_solve(model, solver_name)
     elseif is_available[1] && spec_nls[1] # NLS
-      nls = OptimizationProblems.ADNLPProblems.arglina(; use_nls=true)
+      nls = OptimizationProblems.ADNLPProblems.arglina(; use_nls = true)
       test_in_place_solve(nls, solver_name)
     elseif is_available[1] # RipQP
       nlp_qm = QuadraticModel(nlp, nlp.meta.x0)
@@ -85,15 +85,15 @@ end
 include("qp_tests.jl")
 
 @testset "Test `Float32`" begin
-  nlp = OptimizationProblems.ADNLPProblems.genrose(; type=Float32)
+  nlp = OptimizationProblems.ADNLPProblems.genrose(; type = Float32)
   atol, rtol = √eps(Float32), √eps(Float32)
   for solver in eachrow(JSOSuite.select_optimizers(nlp))
     if solver.nonlinear_obj
-      minimize(solver.name, nlp; verbose=0, atol=atol, rtol=rtol)
+      minimize(solver.name, nlp; verbose = 0, atol = atol, rtol = rtol)
       @test true
     else
       nlp_qm = QuadraticModel(nlp, nlp.meta.x0)
-      minimize(solver.name, nlp_qm; verbose=0, atol=atol, rtol=rtol)
+      minimize(solver.name, nlp_qm; verbose = 0, atol = atol, rtol = rtol)
       @test true
[…]
