+++ previews/PR163/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"60-speed-up/#speed-up","page":"Speed up Solvers Tips","title":"Speed up Solvers Tips","text":"","category":"section"},{"location":"60-speed-up/","page":"Speed up Solvers Tips","title":"Speed up Solvers Tips","text":"The following contains a list of tips to speed up the solver selection and usage.","category":"page"},{"location":"60-speed-up/#Derivatives","page":"Speed up Solvers Tips","title":"Derivatives","text":"","category":"section"},{"location":"60-speed-up/","page":"Speed up Solvers Tips","title":"Speed up Solvers Tips","text":"The optimizers available in JSOSuite.jl are all using first and sometines second-order derivatives. There are mainly three categories:","category":"page"},{"location":"60-speed-up/","page":"Speed up Solvers Tips","title":"Speed up Solvers Tips","text":"1st order methods use only gradient information;\n1st order quasi-Newton methods require only gradient information, and uses it to build an approximation of the Hessian;\n2nd order methods: Those are using gradients and Hessian information.\n2nd order methods matrix-free: Those are optimizers using Hessian information, but without ever forming the matrix, so only matrix-vector products are computed.","category":"page"},{"location":"60-speed-up/","page":"Speed up Solvers Tips","title":"Speed up Solvers Tips","text":"The latter is usually a good tradeoff for very large problems.","category":"page"},{"location":"60-speed-up/","page":"Speed up Solvers Tips","title":"Speed up Solvers Tips","text":"using JSOSuite\nf = x -> 100 * (x[2] - x[1]^2)^2 + (x[1] - 1)^2\nx0 = [-1.2; 1.0]\nstats = minimize(f, x0, verbose = 0, highest_derivative_available = 1)\nstats","category":"page"},{"location":"60-speed-up/","page":"Speed up Solvers Tips","title":"Speed up Solvers Tips","text":"This classification is straightforwardly extended to handling constraints with the Jacobian matrix explicitly of via matrix-vector products.","category":"page"},{"location":"60-speed-up/#Find-a-better-initial-guess","page":"Speed up Solvers Tips","title":"Find a better initial guess","text":"","category":"section"},{"location":"60-speed-up/","page":"Speed up Solvers Tips","title":"Speed up Solvers Tips","text":"The majority of derivative-based optimizers are local methods whose performance are dependent of the initial guess. This usually relies on specific knowledge of the problem.","category":"page"},{"location":"60-speed-up/","page":"Speed up Solvers Tips","title":"Speed up Solvers Tips","text":"The function feasible_point computes a point satisfying the constraints of the problem that can be used as an initial guess. An alternative is to solve a simpler version of the problem and reuse the solution as an initial guess for the more complex one.","category":"page"},{"location":"60-speed-up/#Use-the-structure-of-the-problem","page":"Speed up Solvers Tips","title":"Use the structure of the problem","text":"","category":"section"},{"location":"60-speed-up/","page":"Speed up Solvers Tips","title":"Speed up Solvers Tips","text":"If the problem has linear constraints, then it is efficient to specify it at the modeling stage to avoid having them treated like nonlinear ones. Some of the optimizers will also exploit this information.","category":"page"},{"location":"60-speed-up/","page":"Speed up Solv
[…]
