### 2024-06-17 #161 @mzagorowska
I am done as well.

The code of conduct and conflict of interests links in the checklist don't work for me, so no ticks. Other missing ticks: I didn't find any comparison with other similar works in the paper. The [actual documentation ](https://docs.juliahub.com/General/JSOSuite/0.1.1/) does contain [benchmarking ](https://docs.juliahub.com/General/JSOSuite/0.1.1/benchmark/) together with all the functionality descriptions and examples, but the link isn't in the paper. A suggestion - maybe include a brief description of the benchmarks & results in the paper, and provide a link to the full analysis?
