### 2024-06-17 dbee44b
Clean BIB file (#169)
