+++ paper/paper.tex
@@ -27,22 +27,16 @@ This new package caters to practitioners as it does not require an understanding
 \end{equation}
 where \(f:\mathbb{R}^n\to\mathbb{R}\) and  \(c:\mathbb{R}^n\to\mathbb{R}^m\) are continuously differentiable, with \(c_L \in \left(\mathbb{R} \cup \{-\infty\} \right)^{m}\),  \(c_U \in \left(\mathbb{R} \cup \{+\infty\} \right)^{m}\),  \(\ell \in \left(\mathbb{R} \cup \{-\infty\} \right)^n\), and  \(u \in \left(\mathbb{R} \cup \{+\infty\} \right)^n\).
 Bounds on the variables appear separately from other types of constraints because numerical methods often treat them differently.
-%Equality constraints are modeled by setting components  \(-\infty < c_{L,i}=c_{U,i} < +\infty\) in~\eqref{eq:nlp}.
 
 \section{Statement of Need}
 
-\texttt{JSOSuite.jl} is part of the JuliaSmoothOptimizers (JSO) ecosystem, % \cite{jso}.
-an academic organization that offers a collection of Julia packages for nonlinear optimization software development, testing, and benchmarking.
-
+\texttt{JSOSuite.jl} is part of the JuliaSmoothOptimizers (JSO) ecosystem, an academic organization that offers a collection of Julia packages for nonlinear optimization software development, testing, and benchmarking.
 
 JSO provides a general API for solvers to interact with models by providing flexible data types to represent the objective and constraint functions, evaluate their derivatives, and provide essentially any information that a solver might request from a model.
 %
-\texttt{NLPModels.jl} \cite{orban-siqueira-nlpmodels-2020} is the core package that defines the abstract type \texttt{AbstractNLPModel} and introduces the API, including in-place and out-of-place evaluation of the objective gradient, sparse Jacobian and Hessian matrices as well as operators %\cite{orban-siqueira-linearoperators-2020}
-for matrix-free implementations.
+\texttt{NLPModels.jl} \cite{orban-siqueira-nlpmodels-2020} is the core package that defines the abstract type \texttt{AbstractNLPModel} and introduces the API, including in-place and out-of-place evaluation of the objective gradient, sparse Jacobian and Hessian matrices as well as operators for matrix-free implementations.
 %
-The user can hand-code derivatives %\cite{Siqueira_ManualNLPModels_NLPModels_where_2022}
-, use automatic differentiation \cite{orban-siqueira-nlpmodels-2021}, or use JSO interfaces to optimization modeling languages such as AMPL \cite{fourer2003ampl}, CUTEst \cite{cutest}, or JuMP \cite{jump}.
-% \cite{orban-siqueira-amplnlreader-2020,orban-siqueira-cutest-2020,montoison-orban-siquiera-nlpmodelsjump-2020}
+The user can hand-code derivatives, use automatic differentiation \cite{orban-siqueira-nlpmodels-2021}, or use JSO interfaces to optimization modeling languages such as AMPL \cite{fourer2003ampl}, CUTEst \cite{cutest}, or JuMP \cite{jump}.
 %
 Hence, solvers can be designed to rely on the API's behavior independently of the problem's origin. 
 
@@ -54,8 +48,6 @@ There are a growing number of JSO-compliant solvers accessible from registered J
 Most of those solvers also have variants for nonlinear least squares.
 A solver for convex quadratic programs is in \texttt{RipQP.jl} \cite{orban-leconte-ripqp-2020}.
[…]
