+++ paper/ref.bib
@@ -31,18 +31,6 @@ version = {0.20.0},
 year = {2023}
 }
 
-@article{montoison-orban-krylov-2020,
-author = {Montoison, Alexis and Orban, Dominique},
-doi = {10.21105/joss.05187},
-journal = {Journal of Open Source Software},
-number = {89},
-pages = {5187},
-title = {{Krylov.jl: A Julia basket of hand-picked Krylov methods}},
-url = {https://joss.theoj.org/papers/10.21105/joss.05187},
-volume = {8},
-year = {2023}
-}
-
 @software{orban-siqueira-nlpmodels-2021,
 author = {Migot, Tangi and Montoison, Alexis and Orban, Dominique and Soares Siqueira, Abel and contributors},
 license = {MPL-2.0},
@@ -51,15 +39,6 @@ url = {https://github.com/JuliaSmoothOptimizers/ADNLPModels.jl},
 DOI = {10.5281/zenodo.4605982},
 version = {0.7.0},
 year = {2023}
-}
-
-@software{Siqueira_ManualNLPModels_NLPModels_where_2022,
-author = {Siqueira, Abel Soares},
-doi = {10.5281/zenodo.6940304},
-url = {https://github.com/JuliaSmoothOptimizers/ManualNLPModels.jl},
-title = {{ManualNLPModels: NLPModels where all functions are defined by the user}},
-version = {0.1.1},
-year = {2022}
 }
 
 @article{bezanson2017julia,
@@ -74,44 +53,10 @@ year = {2022}
   doi={10.1137/141000671}
 }
 
-@article{lubin2015computing,
-  title={Computing in operations research using {J}ulia},
-  author={Lubin, Miles and Dunning, Iain},
-  journal={INFORMS Journal on Computing},
-  volume={27},
-  number={2},
-  pages={238--248},
-  year={2015},
-  publisher={INFORMS},
-  doi={10.1287/ijoc.2014.0623}
-}
-
-@software{orban-siqueira-nlpmodelsknitro-2020,
-author = {Migot, Tangi and Orban, Dominique and Soares Siqueira, Abel and contributors},
-license = {MPL-2.0},
-title = {{NLPModelsKnitro.jl: A thin KNITRO wrapper for NLPModels}},
-url = {https://github.com/JuliaSmoothOptimizers/NLPModelsKnitro.jl},
-DOI = {10.5281/zenodo.3994983},
-version = {0.8.0},
-year = {2023}
-}
-
-@software{orban-siqueira-nlpmodelsipopt-2020,
-author = {Orban, Dominique and Soares Siqueira, Abel and contributors},
-license = {MPL-2.0},
-title = {{NLPModelsIpopt.jl: A thin IPOPT wrapper for NLPModels}},
-url = {https://github.com/JuliaSmoothOptimizers/NLPModelsIpopt.jl},
-DOI = {10.5281/zenodo.2629034},
-version = {0.10.1},
-year = {2023}
-}
-
 @Inbook{byrd2006k,
 author="Byrd, Richard H.
 and Nocedal, Jorge
 and Waltz, Richard A.",
-editor="Di Pillo, G.
-and Roma, M.",
 title="Knitro: An integrated package for nonlinear optimization",
 bookTitle="Large-Scale Nonlinear Optimization",
 year="2006",
@@ -119,14 +64,6 @@ publisher="Springer US",
 address="Boston, MA",
 pages="35--59",
 doi="10.1007/0-387-30065-1",
-}
-
-@software{montoison-orban-siquiera-nlpmodelsjump-2020,
-  author = {Alexis Montoison and Dominique Orban and Abel Soares Siqueira and contributors},
[…]
