### 2024-06-17 #161 @odow
Hmm. The links should probably be:

 * https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy
 * https://proceedings.juliacon.org/about#code_of_conduct

@matbesancon @lucaferranti is this known?
