### 2024-06-18 #161 @lucaferranti
Good catch! Apparently the checklist template got the wrong links at some point. I opened a PR fixing it. Meanwhile, the links given by Oscar are good to refer too to tick the corresponding boxes
