+++ lychee.toml
@@ -1,9 +1,5 @@
 exclude = [
-  "@ref",
-  "^https://github.com/.*/releases/tag/v.*$",
-  "^https://doi.org/FIXME$",
-  "^https://JuliaSmoothOptimizers.github.io/JSOSuite.jl/stable$",
-  "zenodo.org/badge/DOI/FIXME$"
+  "@ref"
 ]
 
 exclude_path = [
