### 2024-06-21 74ca17e
Fix some issues pointed out by @tmigot
