### 2024-06-22 75d7c2a
CompatHelper: bump compat for NLPModelsJuMP to 0.13, (keep existing compat)
