+++ Project.toml
@@ -31,7 +31,7 @@ JuMP = "1"
 LLSModels = "0.3.6"
 LinearOperators = "2"
 NLPModels = "0.19, 0.20, 0.21"
-NLPModelsJuMP = "0.11, 0.12"
+NLPModelsJuMP = "0.11, 0.12, 0.13"
 NLPModelsModifiers = "0.6.2, 0.7"
 Percival = "0.6, 0.7"
 QuadraticModels = "0.9.2"
