### 2024-06-28 93385a0
Update to COPIERTemplate 0.6.1
