+++ .github/ISSUE_TEMPLATE/30-usage.yml
@@ -5,13 +5,13 @@ body:
   - type: markdown
     attributes:
       value: |
-        Thanks for taking the time to fill out this bug report!
+        Thanks for taking the time to fill out this question!
 
         Please, before submitting, make sure that:
 
         - You have checked the [documentation](https://JuliaSmoothOptimizers.github.io/JSOSuite.jl) and haven't found enough information
         - There is not an [existing issue](https://github.com/JuliaSmoothOptimizers/JSOSuite.jl/issues) with the same question
-        - You have read the [contributing guide](https://github.com/JuliaSmoothOptimizers/JSOSuite.jl/blob/main/docs/src/90-contributing.md/)
+        - You have read the [contributing guide](https://JuliaSmoothOptimizers.github.io/JSOSuite.jl/dev/90-contributing/)
         - You are following the [code of conduct](https://github.com/JuliaSmoothOptimizers/JSOSuite.jl/blob/main/CODE_OF_CONDUCT.md)
 
         The form below should help you in filling out this issue.
