### 2024-06-28 a3b2e96
Update to COPIERTemplate 0.6.1
