+++ .github/ISSUE_TEMPLATE/99-general.yml
@@ -4,12 +4,12 @@ body:
   - type: markdown
     attributes:
       value: |
-        Thanks for taking the time to fill out this bug report!
+        Thanks for taking the time to fill out this issue!
 
         Please, before submitting, make sure that:
 
         - There is not an [existing issue](https://github.com/JuliaSmoothOptimizers/JSOSuite.jl/issues) with the same question
-        - You have read the [contributing guide](https://github.com/JuliaSmoothOptimizers/JSOSuite.jl/blob/main/docs/src/90-contributing.md/)
+        - You have read the [contributing guide](https://JuliaSmoothOptimizers.github.io/JSOSuite.jl/dev/90-contributing/)
         - You are following the [code of conduct](https://github.com/JuliaSmoothOptimizers/JSOSuite.jl/blob/main/CODE_OF_CONDUCT.md)
 
         The form below should help you in filling out this issue.
