+++ .github/ISSUE_TEMPLATE/20-feature-request.yml
@@ -4,12 +4,12 @@ body:
   - type: markdown
     attributes:
       value: |
-        Thanks for taking the time to fill out this bug report!
+        Thanks for taking the time to fill out this feature request!
 
         Please, before submitting, make sure that:
 
         - There is not an [existing issue](https://github.com/JuliaSmoothOptimizers/JSOSuite.jl/issues) with the same question
-        - You have read the [contributing guide](https://github.com/JuliaSmoothOptimizers/JSOSuite.jl/blob/main/docs/src/90-contributing.md/)
+        - You have read the [contributing guide](https://JuliaSmoothOptimizers.github.io/JSOSuite.jl/dev/90-contributing/)
         - You are following the [code of conduct](https://github.com/JuliaSmoothOptimizers/JSOSuite.jl/blob/main/CODE_OF_CONDUCT.md)
 
         The form below should help you in filling out this issue.
@@ -20,6 +20,11 @@ body:
       description: Describe the requested feature
     validations:
       required: true
+  - type: textarea
+    id: validation
+    attributes:
+      label: Validation and testing
+      description: How could we verify that the new feature works? What kind of tests can be done?
   - type: textarea
     id: motivation
     attributes:
@@ -33,5 +38,5 @@ body:
   - type: textarea
     id: can-help
     attributes:
-      label: Can you help
+      label: Can you help?
       description: Can you help developing this feature?
