+++ Project.toml
@@ -1,7 +1,7 @@
 name = "JSOSuite"
 uuid = "ed6ae0be-a024-11e9-2788-05dbf8cd15d9"
 authors = ["Tangi Migot <tangi.migot@gmail.com>"]
-version = "0.1.1"
+version = "0.2.0"
 
 [deps]
 ADNLPModels = "54578032-b7ea-4c30-94aa-7cbd1cce6c9a"
