### 2024-07-03 cd3751d
COPIERTemplate has been renamed to BestieTemplate
