+++ paper/paper.yml
@@ -16,7 +16,7 @@ authors:
     orcid: 0000-0003-4451-281X
     affiliation: 2
 affiliations:
-  - name: GERAD and Department of Mathematics and Industrial Engineering, Polytechnique Montréal, Montr\'eal (Qc), Canada, H3T 2A7
+  - name: GERAD and Department of Mathematics and Industrial Engineering, Polytechnique Montréal, Canada
     index: 1
   - name: Netherlands eScience Center, Amsterdam, Netherlands
     index: 2
