### 2024-07-06 fc802ad
Add comments about alternatives (#173)
