+++ paper/ref.bib
@@ -166,22 +166,22 @@ version = {0.2.4},
 year = {2023}
 }
 
-@article{cutest,
-  author = {Gould, Nicholas I. and Orban, Dominique and Toint, Philippe L.},
-  title = {{CUTEst}: A Constrained and Unconstrained Testing Environment with Safe
-           Threads for Mathematical Optimization},
-  journal = {Computational Optimization and Applications},
-  volume = {60},
-  number = {3},
-  year = {2015},
-  pages = {545--557},
-  doi = {10.1007/s10589-014-9687-3},
-}
-
 @article{shin2023accelerating,
   title={Accelerating optimal power flow with {GPU}s: {SIMD} abstraction of nonlinear programs and condensed-space interior-point methods},
   author={Shin, Sungho and Pacaud, Fran{\c{c}}ois and Anitescu, Mihai},
   journal={arXiv preprint arXiv:2307.16830},
   year={2023},
   doi = {10.2139/ssrn.4601442}
+}
+
+@article{legat2021mathoptinterface,
+    title={{MathOptInterface}: a data structure for mathematical optimization problems},
+    author={Legat, Beno{\^\i}t and Dowson, Oscar and Dias Garcia, Joaquim and Lubin, Miles},
+    journal={INFORMS Journal on Computing},
+    year={2021},
+    volume={34},
+    number={2},
+    pages={672--689},
+    doi={10.1287/ijoc.2021.1067},
+    publisher={INFORMS}
 }
