+++ previews/PR173/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"nls/#nls-section","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"","category":"section"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"The nonlinear least squares (NLS) optimization problem is a specific case where the objective function is a sum of squares.","category":"page"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"beginaligned\nmin quad  f(x)=tfrac12F(x)^2_2 \n c_L leq c(x) leq c_U \n c_A leq Ax leq l_A \n ell leq x leq u\nendaligned","category":"page"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"Although the problem can be solved using only  f, knowing  F independently allows the development of more efficient methods.","category":"page"},{"location":"nls/#Model-and-solve-NLS","page":"Nonlinear Least Squares","title":"Model and solve NLS","text":"","category":"section"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"In this tutorial, we consider the following equality-constrained problem","category":"page"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"beginaligned\nmin quad  f(x)=tfrac12(10 * (x2 - x1^2))^2 + tfrac12(x1 - 1)^2 \n 1 leq x1 * x2 leq 1\nendaligned","category":"page"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"where 1 leq x1 x2 leq 1 implies that x1 x2 = 1.","category":"page"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"In the rest of this tutorial, we will see two ways to model this problem exploiting the knowledge of the structure of the problem.","category":"page"},{"location":"nls/#NLS-using-automatic-differentiation","page":"Nonlinear Least Squares","title":"NLS using automatic differentiation","text":"","category":"section"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"Using the package ADNLPModels.jl, the problem can be model as an ADNLSModel which will use automatic-differentiation to compute the derivatives.","category":"page"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"using ADNLPModels, JSOSuite\nF = x -> [10 * (x[2] - x[1]^2); x[1] - 1]\nnres = 2 # size of F(x)\nx0 = [-1.2; 1.0]\nc = x -> [x[1] * x[2]]\nl = [1.]\nnls = ADNLSModel(F, x0, nres, c, l, l, name=\"AD-Rosenbrock\")","category":"page"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"Note that the length of the residual function is given explictly to avoid any superfluous evaluation of this (potentially very large) function.","category":"page"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"stats = minimize(nls)","category":"page"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"JSOSuite.jl uses by default automatic differentiation, so the following code would be equivalent:","category":"page"},{"location":"nls/","page":"Nonlinear Least Squares","title":"Nonlinear Least Squares","text":"stats = minimize(F, x0, nres, c, l, l)","category":"page"},{"location":"nls/","page":"Nonlinear Least Squares","tit
[…]
