+++ README.md
@@ -8,7 +8,7 @@
 [![Docs workflow Status](https://github.com/JuliaSmoothOptimizers/JSOSuite.jl/actions/workflows/Docs.yml/badge.svg?branch=main)](https://github.com/JuliaSmoothOptimizers/JSOSuite.jl/actions/workflows/Docs.yml?query=branch%3Amain)
 [![Build Status](https://api.cirrus-ci.com/github/JuliaSmoothOptimizers/JSOSuite.jl.svg)](https://cirrus-ci.com/github/JuliaSmoothOptimizers/JSOSuite.jl)
 [![Coverage](https://codecov.io/gh/JuliaSmoothOptimizers/JSOSuite.jl/branch/main/graph/badge.svg)](https://codecov.io/gh/JuliaSmoothOptimizers/JSOSuite.jl)
-[![DOI](https://zenodo.org/badge/DOI/FIXME)](https://doi.org/FIXME)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.12588054.svg)](https://doi.org/10.5281/zenodo.12588054)
 [![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-2.1-4baaaa.svg)](CODE_OF_CONDUCT.md)
 [![All Contributors](https://img.shields.io/github/all-contributors/JuliaSmoothOptimizers/JSOSuite.jl?labelColor=5e1ec7&color=c0ffee&style=flat-square))](#contributors)
 
