+++ .cirrus.yml
@@ -1,31 +1,12 @@
+freebsd_instance:
+  image_family: freebsd-13-3
 task:
-  matrix:
-    - name: FreeBSD
-      freebsd_instance:
-        image_family: freebsd-13-2
-      env:
-        matrix:
-          - JULIA_VERSION: 1.6
-          - JULIA_VERSION: 1
-    - name: MacOS M1
-      macos_instance:
-        image: ghcr.io/cirruslabs/macos-monterey-base:latest
-      env:
-        - JULIA_VERSION: 1
-  install_script: |
-    URL="https://raw.githubusercontent.com/ararslan/CirrusCI.jl/master/bin/install.sh"
-    set -x
-    if [ "$(uname -s)" = "Linux" ] && command -v apt; then
-        apt update
-        apt install -y curl
-    fi
-    if command -v curl; then
-        sh -c "$(curl ${URL})"
-    elif command -v wget; then
-        sh -c "$(wget ${URL} -q -O-)"
-    elif command -v fetch; then
-        sh -c "$(fetch ${URL} -o -)"
-    fi
+  name: FreeBSD
+  env:
+    matrix:
+      - JULIA_VERSION: 1
+  install_script:
+    - sh -c "$(fetch https://raw.githubusercontent.com/ararslan/CirrusCI.jl/master/bin/install.sh -o -)"
   build_script:
     - cirrusjl build
   test_script:
