### 2024-07-07 eea5852
chore: 🤖 Apply BestieTemplate v0.7.2 (#163)
Co-authored-by: Tangi Migot <tangi.migot@gmail.com>
