+++ test/runtests.jl
@@ -73,7 +73,7 @@ end
       test_in_place_solve(nlp, solver_name)
       test_in_place_solve(model, solver_name)
     elseif is_available[1] && spec_nls[1] # NLS
-      nls = OptimizationProblems.ADNLPProblems.arglina(use_nls = true)
+      nls = OptimizationProblems.ADNLPProblems.arglina(; use_nls = true)
       test_in_place_solve(nls, solver_name)
     elseif is_available[1] # RipQP
       nlp_qm = QuadraticModel(nlp, nlp.meta.x0)
@@ -85,15 +85,15 @@ end
 include("qp_tests.jl")
 
 @testset "Test `Float32`" begin
-  nlp = OptimizationProblems.ADNLPProblems.genrose(type = Float32)
+  nlp = OptimizationProblems.ADNLPProblems.genrose(; type = Float32)
   atol, rtol = √eps(Float32), √eps(Float32)
   for solver in eachrow(JSOSuite.select_optimizers(nlp))
     if solver.nonlinear_obj
-      minimize(solver.name, nlp, verbose = 0, atol = atol, rtol = rtol)
+      minimize(solver.name, nlp; verbose = 0, atol = atol, rtol = rtol)
       @test true
     else
       nlp_qm = QuadraticModel(nlp, nlp.meta.x0)
-      minimize(solver.name, nlp_qm, verbose = 0, atol = atol, rtol = rtol)
+      minimize(solver.name, nlp_qm; verbose = 0, atol = atol, rtol = rtol)
       @test true
     end
   end
@@ -104,47 +104,47 @@ end
   jum = MathOptNLPModel(model)
   @test JSOSuite.select_optimizers(model) == JSOSuite.select_optimizers(jum)
   for solver in eachrow(JSOSuite.select_optimizers(model))
-    minimize(solver.name, model, verbose = 0)
+    minimize(solver.name, model; verbose = 0)
     @test true
   end
 end
 
 @testset "Benchmark on unconstrained problems" begin
   ad_problems = [
-    OptimizationProblems.ADNLPProblems.eval(Meta.parse(problem))() for problem ∈
+    OptimizationProblems.ADNLPProblems.eval(Meta.parse(problem))() for problem in
     first(meta[(5 .<= meta.nvar .<= 10) .& (meta.ncon .== 0) .& (.!meta.has_bounds), :name], 5)
   ]
   select = JSOSuite.optimizers[
     JSOSuite.optimizers.can_solve_nlp .& JSOSuite.optimizers.is_available,
     :name,
   ]
-  stats = bmark_solvers(ad_problems, select, atol = 1e-3, max_time = 10.0, verbose = 0)
+  stats = bmark_solvers(ad_problems, select; atol = 1e-3, max_time = 10.0, verbose = 0)
   @test true # just test that it runs
 end
 
 @testset "Basic solve tests" begin
   f = x -> 100 * (x[2] - x[1]^2)^2 + (x[1] - 1)^2
-  stats = minimize(f, [-1.2; 1.0], verbose = 0)
+  stats = minimize(f, [-1.2; 1.0]; verbose = 0)
   @test stats.status_reliable && (stats.status == :first_order)
 
-  stats = minimize("DCISolver", f, [-1.2; 1.0], verbose = 0)
+  stats = minimize("DCISolver", f, [-1.2; 1.0]; verbose = 0)
   @test stats.status_reliable && (stats.status == :first_order)
 
   F = x -> [10 * (x[2] - x[1]^2); x[1] - 1]
-  stats = minimize(F, [-1.2; 1.0], 2, verbose = 0)
+  stats = minimize(F, [-1.2; 1.0], 2; verbose = 0)
   @test stats.status_reliable && (stats.status == :first_order)
 
[…]
