+++ docs/Project.toml
@@ -7,6 +7,7 @@ GR = "28b8d3ca-fb5f-59d9-8090-bfdbd6d07a71"
 JSOSolvers = "10dff2fc-5484-5881-a0e0-c90441020f8a"
 JSOSuite = "ed6ae0be-a024-11e9-2788-05dbf8cd15d9"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
+LiveServer = "16fef848-5104-11e9-1b77-fb7a48bbb589"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 NLPModels = "a4795742-8479-5a88-8948-cc11e1c8c1a6"
 NLPModelsIpopt = "f4238b75-b362-5c4c-b852-0801c9a21d71"

[…]
