### 2024-07-12 7a641ec
Detect NLS pattern
