+++ test/Project.toml
@@ -3,6 +3,7 @@ ADNLPModels = "54578032-b7ea-4c30-94aa-7cbd1cce6c9a"
 CaNNOLeS = "5a1c9e79-9c58-5ec0-afc4-3298fdea2875"
 DCISolver = "bee2e536-65f6-11e9-3844-e5bb4c9c55c9"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+ExpressionTreeForge = "93090adf-0e31-445f-8c8f-44d91f61d7ad"
 FletcherPenaltySolver = "e59f0261-166d-4fee-8bf3-5e50457de5db"
 JSOSolvers = "10dff2fc-5484-5881-a0e0-c90441020f8a"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
@@ -32,6 +33,7 @@ ADNLPModels = "0.8.4"
 CaNNOLeS = "0.7"
 DCISolver = "0.4"
 DataFrames = "1"
+ExpressionTreeForge = "^0.1.7"
 FletcherPenaltySolver = "0.2"
 JSOSolvers = "0.10, 0.11"
 JuMP = "1"
