### 2024-07-12 a1bf5d8
Detect NLS pattern
