+++ paper/paper.tex
@@ -28,7 +28,7 @@ This new package caters to practitioners as it does not require an understanding
 where \(f:\mathbb{R}^n\to\mathbb{R}\) and  \(c:\mathbb{R}^n\to\mathbb{R}^m\) are continuously differentiable, with \(c_L \in \left(\mathbb{R} \cup \{-\infty\} \right)^{m}\),  \(c_U \in \left(\mathbb{R} \cup \{+\infty\} \right)^{m}\),  \(\ell \in \left(\mathbb{R} \cup \{-\infty\} \right)^n\), and  \(u \in \left(\mathbb{R} \cup \{+\infty\} \right)^n\).
 Bounds on the variables appear separately from other types of constraints because numerical methods often treat them differently.
 
-\section{Statement of Need}
+\section{Statement of need}
 
 \texttt{JSOSuite.jl} is part of the JuliaSmoothOptimizers (JSO) ecosystem, an academic organization that offers a collection of Julia packages for nonlinear optimization software development, testing, and benchmarking.
 
@@ -109,7 +109,7 @@ Combined with \texttt{SolverBenchmark.jl} \cite{orban-siqueira-solverbenchmark-2
 Additionally, the package implements strategies that are classical for continuous optimization solvers such as methods to find a feasible initial guess or run a multi-start strategy.
 The documentation of the package \url{https://jso.dev/JSOSuite.jl} contains further examples of these functionalities.
 
-\section{Concluding Remarks}
+\section{Concluding remarks}
 
 Julia’s JIT compiler is ideal for efficient scientific computing and optimization software, making it a natural choice for developing new solvers.
 %
