### 2024-07-14 #161 @odow
I opened a small tweak: https://github.com/JuliaSmoothOptimizers/JSOSuite.jl/pull/175
