### 2024-07-22 d636f29
chore: :robot: pre-commit update (#177)
Co-authored-by: abelsiqueira <1068752+abelsiqueira@users.noreply.github.com>
