+++ .pre-commit-config.yaml
@@ -44,6 +44,6 @@ repos:
     hooks:
       - id: yamllint
   - repo: https://github.com/domluna/JuliaFormatter.jl
-    rev: v1.0.57
+    rev: v1.0.58
     hooks:
       - id: julia-formatter
