+++ test/Project.toml
@@ -28,7 +28,7 @@ SparseMatricesCOO = "fa32481b-f100-4b48-8dc8-c62f61b13870"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
-ADNLPModels = "0.4, 0.5, 0.6, 0.7"
+ADNLPModels = "0.8.4"
 CaNNOLeS = "0.7"
 DCISolver = "0.4"
 DataFrames = "1"
