### 2024-08-18 b03f931
Update src/parse-functions.jl
Co-authored-by: Paul Raynaud <paul.raynaud66@hotmail.fr>
