+++ src/parse-functions.jl
@@ -16,7 +16,6 @@ function isnls(nlp)
   expr_tree = ExpressionTreeForge.get_expression_tree(nlp)
   F_expr = extract_element_functions(expr_tree)
   test_square(expr) = expr.field == ExpressionTreeForge.M_power_operator.Power_operator{Int}(2)
-  myand(a::Bool,b::Bool) = a && b
-  is_nls = mapreduce(test_square, myand, F_expr) :: Bool
+  is_nls = mapreduce(test_square, &, F_expr) :: Bool
   return is_nls 
 end 
