### 2024-08-26 #161 @odow
@tmigot can you upload a version of JSO to https://zenodo.org to get a DOI please?
