### 2024-09-01 #161 @tmigot
> @tmigot can you upload a version of JSO to https://zenodo.org to get a DOI please?

It is already done, here: https://zenodo.org/records/12588054
