### 2024-09-01 6f7a5c4
Add `ExpressionTreeForge.extract_element_functions` in parse-function.jl (#181)
* Add `ExpressionTreeForge.extract_element_functions` in parse-function.jl

* Edit parse-function-test.jl

Allow all tests to run, i.e. does not stop at the first error.
