+++ test/parse-function-test.jl
@@ -1,5 +1,6 @@
-problems_names = meta[!,:name][1:50] # we test only the 50 first problems
-problems_symbols = map(name -> OptimizationProblems.ADNLPProblems.eval(Meta.parse(name)), problems_names)
+problems_names = meta[!, :name][1:50] # we test only the 50 first problems
+problems_symbols =
+  map(name -> OptimizationProblems.ADNLPProblems.eval(Meta.parse(name)), problems_names)
 
 @testset "NLS from NLP" begin
   result_list_isnls = Bool[]
@@ -12,11 +13,11 @@ problems_symbols = map(name -> OptimizationProblems.ADNLPProblems.eval(Meta.pars
         @test detect_nls == is_nls
         @debug "Least squares objective, $index : $problem"
         # detect_nls != is_nls && (@debug "$index : $problem")
-      else 
+      else
         @debug "Not least squares objective, $index : $problem"
       end
-    catch 
+    catch
       @debug "the $index-th $(problems_names[index]) problem is unsupported."
-    end 
+    end
   end
 end
