### 2024-09-01 a5e8009
Detect NLS pattern (#174)
* Detect NLS pattern

* Update src/parse-functions.jl

Co-authored-by: Paul Raynaud <paul.raynaud66@hotmail.fr>

* Add `ExpressionTreeForge.extract_element_functions` in parse-function.jl (#181)

* Add `ExpressionTreeForge.extract_element_functions` in parse-function.jl

* Edit parse-function-test.jl

Allow all tests to run, i.e. does not stop at the first error.

* fix docstring

* apply linter

---------

Co-authored-by: Paul Raynaud <paul.raynaud66@hotmail.fr>
