### 2024-09-02 f5428dd
Update JSOSuite citation (#183)
