+++ CITATION.cff
@@ -39,6 +39,25 @@ identifiers:
   - description: Zenodo archive
     type: doi
     value: 10.5281/zenodo.8246389
+preferred-citation:
+  type: article
+  authors:
+  - family-names: "Migot"
+    given-names: "Tangi"
+    orcid: "https://orcid.org/0000-0001-7729-2513"
+  - family-names: "Orban"
+    given-names: "Dominique"
+    orcid: "https://orcid.org/0000-0002-8017-7687"
+  - family-names: "Soares Siqueira"
+    given-names: "Abel"
+    orcid: "https://orcid.org/0000-0003-4451-281X"
+  doi: "10.21105/jcon.00161"
+  journal: "Proceedings of the JuliaCon Conferences"
+  start: 161 # First page number
+  title: "JSOSuite.jl: Solving continuous optimization problems with JuliaSmoothOptimizers"
+  issue: 63
+  volume: 6
+  year: 2024
 repository-code: 'https://github.com/JuliaSmoothOptimizers/JSOSuite.jl'
 version: 0.1.0
 date-released: '2023-08-14'
